Create RevlexOS-Install-v-1.0
The original Installer for RevlexOS 1.0

If you decide to install this OS just press enter on the part where it asks for a key. I didn't think that product keys were such a great idea for RevlexOS 1.1 so I went back here and removed it from RevlexOS 1.0.

I do realize that this installer is horribly written, as this was version 1.0 and I was new to ComputerCraft. Same goes for RevlexOS 1.0, I was new to ComputerCraft, and I realize it was poorly written, but here you go anyway.